requirements.txt
fastapi
uvicorn
langchain
google-generativeai
pydantic
numpy
pandas
scikit-learn
plotly
matplotlib
python-multipart
aiohttp
requests
boto3
minio
python-dotenv
